buildPatchFromRevert
